backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.7.0
build==1.3.0
cachetools==6.2.2
certifi==2023.11.17
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
cloud-init==25.1.4
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
crewai==1.7.0
crewai-tools==1.7.0
cryptography==41.0.7
dbus-python==1.3.2
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
distro-info==1.7+build1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.124.4
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.4
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
identify==2.6.15
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==22.10.0
instructor==1.12.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.32
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lancedb==0.5.7
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mcp==1.16.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
netifaces==0.11.0
networkx==3.6.1
nodeenv==1.9.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.2
onnxruntime==1.23.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.1
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
propcache==0.4.1
protobuf==5.29.5
py==1.11.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.2.8
pybase64==1.4.3
pycurl==7.45.3
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.9.0
pylance==0.9.18
PyMuPDF==1.26.7
PyMySQL==1.1.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.16.2
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semver==3.0.4
sentence-transformers==5.2.0
service-identity==24.1.0
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
sqlmodel==0.0.27
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
systemd-python==235
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.22.1
tomli==2.0.2
tomli_w==1.1.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
Twisted==24.3.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
uv==0.9.17
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
wadllib==1.3.6
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zope.interface==6.1
